requirements.txt
Django==1.11.6
selenium==3.6.0
behave-django==0.5.0
